Update 2018-05-03-script-draft.md
copied over the version of the thesis outline shared in lab meeting today. 

notes from today: 
 - more DRAMA in the plot!
 - the plot should come to conclusion:
    Neuroscience is extremely diverse in the expertise considered necessary to do well, and also oddly familiar to everyone. Min journals a lot, and cries every time she does it. Without distributed intelligent thinking systems, like EveryMind.online, it becomes easy to be manipulated through aggressive datamining in the name of economic gain. 

Asoca should experience something that makes her realise a superpower, she needs to have an enlightenment moment. 

There's tremendous power if we could coordinate brains in a different way. 

Everyone is a neuroscientist but they just need the tools for understanding their own minds more explicitly. 

 - email Ed Bracey about Arts College students  ---> DONE! HA! :D

conclusion: propose how to re-organise neuroscience to be distributed, to be the "anti-thesis" of the patron model (? ask Adam what he meant, I'm pretty sure I get it though). We don't need the traditional structures anymore, in fact they are choking us. we need a completely different organization in the world where everything is connected and data is mined for economic gain. It's ok to change EVERYTHING about Neuroscience, as long as we are willing to talk about all of our assumptions explicitly! our goals, our incentives, etc. 

Stories are about individuals who fight against themselves and realise who they are, and what they really want. First you don't know who you are and you struggle, then you know who you are and everything is fine. 

Break the concepts of good and bad by making characters live in the grey areas, the uncomfortable areas, the areas that make you feel awkward. 

TV show "Dead wood" 

a little more horror (from Lory hahahaha)

cuttlefish as a motif: they masterfully coordinate millions of distributed elements, how do they coordinate this? lol make up some kind of network math?! talk to mathematician friends from MIT, put out a call for math consultation. How to apply this insight to society and distributed human systems?

The final distillation is that neuroscience needs a model for distributed science in the wild. 

eLife paper: neuroscience lab tools need to be in the public spaces, everyone needs the tools to science on their own, everywhere. 
 - talk about the practical experience of building a "lab outside The Lab". 
 - a new model for how human society learns more about itself. 
 - "World as Neuroscience Lab, Everyone as Neuroscientist"
 - neuroscience needs to figure out how to NOT NEED ACADEMIA

Graphic novel should propose a solution to the post-truth world. 

Main character needs a superpower, some little trick that let's her sidestep the conventional ways of doing science. She needs to have some unique power CAPOEIRA POWERS MANDINGA MANDINGUEIROS OOOOAAAAAAAAAA! to understand herself very intuitively, and she figures out how to know this explicitly. YAAAAASSSSS wait so this is Asoca, because she's the one who plays capoeira, so she's also super disciplined in how to move her body. 

The struggle to become a top 10 student should be the vehicle to talk about the "transfer function" assumptions of brains. Then Min realises that what's really important is to figure out what she wants. 

Asoca's family have all learned how to register in the SPACE CONSORTIUM's intelligence tests as mentally docile and of below average intelligence, so that they are not dosed with government-mandated mood stabilizers and job specification neural mods. 

The ending should be one clear point. What is the takeaway? What was the point of my PhD?

Disclaimer page. For me the process was the most important lesson I learned during my PhD, so in order to convey to you the necessary psychological context, here's a story, a science fiction, made-up graphic novel. 

eLife paper should be a recipe for doing Field Neuroscience. Write it as the appendix to the comic, and my target audience should be people reading the paper after reading the graphic novel. We need to equip people for the "good" tools and ideas from science, and by people we literally mean as many people as possible, so it's gotta be the permaculture way --> existing model for sustainable distributed human intelligent behaviour!!!!!!! 

AWESOME - everything I thought might be important is turning out to be important! Trust means I believe in it enough to incorporate it into my daily practice. Knowing something is having experienced the consequences of this incorporation into my daily practice. AAAAAAAAAAAAAHHHHHHHHHHHH :D :D :D