script restructuring, part 1
the graphic novel now starts with a chapter about young Majhee, grandmother of Min our main character. The World timeline has been updated to include events related to the Biononymous and the rapid advancement of eSyrup technology and applications before the Global Water Crisis.

The main plot of the graphic novel will focus on the principles of the Automata: to know oneself, and how to make a good mess - don't hurt yourself, don't hurt others, and don't hurt other people's stuff.